♻️ Refactor: CardList components
