+++ client/src/ui/CardsList/CardsList.module.scss
@@ -13,7 +13,7 @@
   }
 
   .shrink li:nth-child(6) {
-    display: none;
+    @apply hidden;
   }
 }
