+++ client/src/ui/CardsList/CardsList.tsx
@@ -1,6 +1,6 @@
 import styles from './CardsList.module.scss';
-import Card from '../../ui/Card/Card.tsx';
-import Heading2 from '../../ui/Typography/Heading2.tsx';
+import Card from '../Card/Card.tsx';
+import Heading2 from '../Typography/Heading2.tsx';
 
 interface CardDate {
   img: string;
@@ -19,6 +19,7 @@ const CardsList = ({ title, shrink = false, items }: CardsListProps) => {
   return (
     <div className={styles.cardsListContainer}>
       {title && <Heading2>{title}</Heading2>}
+
       <ul className={`${styles.cardsList} ${shrink ? styles.shrink : ''}`}>
         {items && items.map((el) => <Card key={el.name} data={el} />)}
       </ul>
